requirements.txt
flask==2.2.5
gunicorn==21.2.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
Pillow==9.5.0
tensorflow-cpu== 2.20.0
gdown